fix(server): use exiftool decoded values and unify metadata extraction
There was a lot of unnecessary stuff happening which exiftool should have
already taken care of. It's been replaced with the values provided by the
exiftool package and the code is much cleaner as a result. Further, exiftool
can be used to extract metadata from videos which greatly simplifies things.

Fixes: #2786, #2859, #2953